split profilers (#6261)
